Created the HUD and Player, along with PlayerRotation.
I've created the base framework for the HUD within the game and have completed the player prefab by attaching a body (capsule) and arm cannon (cylinder with pivot). I've begun working on the player rotation and have created a blank file for the GameManager to be used in the future. Additionally, I've renamed the player input action asset and have fixed errors associated with it along with renaming and organizing the assets and folder structure within the project.